型キャストの修正とコードの整形

+++ src/2d/Camera.cpp
@@ -77,7 +77,7 @@ bool Camera::init()
             float zEye = m_pGame->getZEye();
             m_zNear = 0.5f;
             m_zFar = zEye + size.y / 2.0f;
-            initPerspective(glm::radians(60.0f), (float)size.x / size.y, m_zNear, m_zFar);
+            initPerspective(glm::radians(60.0f), size.x / size.y, m_zNear, m_zFar);
             glm::vec3 eye(size.x / 2.0f, size.y / 2.0f, zEye);
             glm::vec3 center(size.x / 2.0f, size.y / 2.0f, 0.0f);
             setPosition(eye);
