+++ tests/cpp-tests/AppDelegate.cpp
@@ -33,8 +33,8 @@ bool AppDelegate::applicationDidFinishLaunching()
         }
 
         auto monitorSize = glViewImpl->getMonitorSize();
-        const int windowPosX = static_cast<int>((monitorSize.x / 2.0f) - (windowWidth / 2.0f));
-        const int windowPosY = static_cast<int>((monitorSize.y / 2.0f) - (windowHeight / 2.0f));
+        const int windowPosX = static_cast<int>(monitorSize.x / 2) - static_cast<int>(windowWidth / 2);
+        const int windowPosY = static_cast<int>(monitorSize.y / 2) - static_cast<int>(windowHeight / 2);
         glViewImpl->setWindowPosition(windowPosX, windowPosY);
 
         game->setGLView(glView);
